tss 259 Add time to Activity Report Application Form (#5)
## Ticket

Resolves
[TSS-259](https://linear.app/nava-platform/issue/TSS-259/add-minutes-worked-to-activityreportapplicationform)

## Changes

Add a hours worked field. This field takes input in quarter hour
increments, and does a conversion to minutes before sending it to the
controller. The DB will save input as minutes. We also do some minor
validation, ensuring a minimum of 15 minutes but there is no current
maximum yet.

## Context for reviewers

![2025-08-22 Hours Worked
Field](https://github.com/user-attachments/assets/9cf7fc22-8f7f-41f5-8330-5ff41cc30404)

## Testing

While testing, there are some bugs unrelated to this PR. However, it
does look like there is a slight UI glitch for a brief moment, where the
form gets the data in minutes, and needs to convert to hours. I can see
the transition and it would make for better UI if the server does the
conversion first, rather than the form.

<!-- reporting-app - begin PR environment info -->
## Preview environment for reporting-app
- Service endpoint:
https://p-5-reporting-app-dev-107408755.us-east-1.elb.amazonaws.com
- Deployed commit: b8f4f1ed1d01e1c0072f28687d3352ad0f34b0e4
<!-- reporting-app - end PR environment info -->

<!-- app - begin PR environment info -->
## Preview environment for app
♻️ Environment destroyed ♻️
<!-- app - end PR environment info -->

---------

Co-authored-by: baonguyenNava <baonguyen@Baos-MacBook-Pro.local>