Clean up test failures, linting, and modify minutes validation on form model (#8)
## Ticket

None

## Changes

Resolves linting errors and spec failures.

Minutes validation on form - Allow nil to prevent spec failures
(short-term change). The UI now is less strict, and the user does not
have to fill in the hours for the activity. When the user attempts to
submit, the backend will now validate and not allow the submission. The
user will need to update "Hours Worked" before they can submit
successfully.

## Context for reviewers

> Testing instructions, background context, more in-depth details of the
implementation, and anything else you'd like to call out or ask
reviewers.

## Testing

I did some manual UI testing to confirm the behavior above. Note, that
we don't have an error handling, and we'll need to have the UI
gracefully handle form validation errors.

<!-- reporting-app - begin PR environment info -->
## Preview environment for reporting-app
- Service endpoint:
https://p-8-reporting-app-dev-282258153.us-east-1.elb.amazonaws.com
- Deployed commit: 5c6eb84fbdef16d399d68551b7e7d8315527ca78
<!-- reporting-app - end PR environment info -->