reporting-app: Create Auth::CognitoAdapater::Config and fill with dummy values in tests (#13)
Previously the tests would expose whatever config values are active in
the runtime environment, which is dangerous. It means running the tests
could have side effects/mutate state in real resources. We mock the
Cognito client itself in the tests, which should avoid most of the
issues, but it's still dangerous to expose potentially real config
values to the test suite. It also means the tests are less isolated from
local configuration. So avoid these issues and explicitly pass in dummy
configuration.

The Config object makes the required configuration more explicit for
users, as well as provides a nicer abstraction for mocking/testing the
adapter.

It would be prudent to eventually ensure that no environment variables
are exposed to tests, except specifically configured ones, and have
better env var testing tools.