Add certification case (#61)
## Ticket

Resolves
[TSS-351](https://linear.app/nava-platform/issue/TSS-351/add-certification-case)

## Changes

Add Certification case model and its views on the staff dashboard

## Context for reviewers


https://docs.google.com/document/d/1LSaqLGH_M2uczlNtHBGb5ucBaUOfovhwHIrLq4oMAsI/edit?tab=t.0

Adding Certification Case which will be part of the new Certification
Business Process. This will eliminate activity report case and exemption
case, along with their business processes.

## Testing

I did some basic smoke testing to see that the views on the staff
dashboard rendered.

> Provide evidence that the code works as expected. Explain what was
done for testing and the results of the test plan. Include screenshots,
[GIF demos](https://www.cockos.com/licecap/), shell commands or output
to help show the changes working as expected. ProTip: you can drag and
drop or paste images into this textbox.

<!-- reporting-app - begin PR environment info -->
## Preview environment for reporting-app
- Service endpoint:
https://p-61-reporting-app-dev-1536723452.us-east-1.elb.amazonaws.com
- Deployed commit: d0fd6d0b3870a3cf0a315d1f372f36623430dfbf
<!-- reporting-app - end PR environment info -->