Update ActivityReportApplicationRecord to allow multiple reporting periods (#75)
## Ticket

Resolves TSS-358

## Changes

**ActivityReportApplicationForm**
- Strata migration to add reporting_periods array column
- migration to migrate existing reporting_period data to
reporting_periods column
- Added strata_attribute :reporting_periods, :year_month, array: true
- removed strata_attribute :reporting_period, :date

## Context for reviewers

As part of the updates to the activity report workflow, this work adds a
reporting_periods column to the activity_reports_application_form to
allow for multiple months to be selected when creating an activity
report application form. This commit only includes updates to the
backend, leaving the UI unchanged. This commit will be followed up with
a separate PR to drop the now unused reporting_period column and then
the UI updates to allow selecting multiple months.

## Testing

The test suite was updated to reflect this changes in this commit as
well as manual testing to ensure there were no changes to the user
experience.

<!-- reporting-app - begin PR environment info -->
## Preview environment for reporting-app
- Service endpoint:
https://p-75-reporting-app-dev-490684015.us-east-1.elb.amazonaws.com
- Deployed commit: 6efcbd56953dbbf8f49a546b15b06d24962bf516
<!-- reporting-app - end PR environment info -->