Initial API framework and endpoints (#45)
To avoid having to write the entire API spec by hand, using `oas_rails`
to help a bit. It's not quite expressive enough to be able to just use
the Yard tags inline for the schemas, but does a little bit of the
lifting. Then `openapi_contracts` helps assert in the existing request
tests that the response match the API spec.

The other main alternative approach looks to be rswag. Didn't start with
it as it's limited to OpenAPI 3.0 and a bit disconnected from the
controller code itself. But could be be used if ultimately preferred.

Related to https://linear.app/nava-platform/issue/TSS-283
Related to https://linear.app/nava-platform/issue/TSS-306