Add Caseworker experience: Request for Information (#84)
## Ticket

Resolves TSS-364

## Changes

This commit adds the caseworker workflow for creating a request for
information.

**New Models:**
  - InformationRequest
  - ActivityReportInformationRequest (child class of InformationRequest)
  - ExemptionClaimInformationRequest(child class of InformationRequest)

**New Table**
  - information_requests
    - type (string for STI)
    - application_id (uuid)
    - staff_comment (text)
    - member_comment (text)
    -  due_date (date)


From the task view, a staff user now has the option to request
additional information from the member. A button was added which, when
clicked, takes the user to a request for information form with a text
box for the staff user's comments. Future work will replace the three
action buttons on the task view with the radio button design
([TSS-385](https://linear.app/nava-platform/issue/TSS-385/update-task-form-approvedenyrfi-buttons-to-radio-buttons)).

Creating the information request will redirect the user to the case
view.

## Context for reviewers

## Testing
Creating a request for information for an activity report application
form:

![request_for_information_activity_report](https://github.com/user-attachments/assets/c042f27a-ff09-498d-a89a-332794befe46)


Creating a request for information for and exemption application form:

![request_for_information_exemption_claim](https://github.com/user-attachments/assets/2a644024-15bd-4ad5-8fd9-5a00592ccf13)


<!-- reporting-app - begin PR environment info -->
## Preview environment for reporting-app
- Service endpoint:
https://p-84-reporting-app-dev-1414300962.us-east-1.elb.amazonaws.com
- Deployed commit: 399fdf54b10666331acfdf74f34cceb7295a6075
<!-- reporting-app - end PR environment info -->